separate _gen _jit _benchmark
