+++ blurry_benchmark.hpp
@@ -1,0 +1,13 @@
+// +build ignore.
+
+#include <Halide.h>
+#include <halide_benchmark.h>
+
+#include "blurry.hpp"
+
+#ifndef BLURRY_BENCHMARK_H_
+#define BLURRY_BENCHMARK_H_
+
+int benchmark(char **argv);
+
+#endif // BLURRY_BENCHMARK_H_
