+++ blurry_jit.hpp
@@ -1,0 +1,12 @@
+// +build ignore.
+
+#include <Halide.h>
+
+#include "blurry.hpp"
+
+#ifndef BLURRY_JIT_H_
+#define BLURRY_JIT_H_
+
+int jit_run(char **argv);
+
+#endif // BLURRY_JIT_H_
