+++ cli/genrun/benchmark.go
@@ -7,7 +7,7 @@ import (
 func benchmarkAction(c *cli.Context) error {
 	runtimePath := c.String("runtime")
 	useBenchmarkOpt()
-	generateOutFilePath, err := generate(runtimePath, c.String("file"))
+	generateOutFilePath, err := generateWithBenchmark(runtimePath, c.String("file"))
 	if err != nil {
 		return err
 	}
