+++ cli/genrun/convert_from_yuv.go
@@ -6,7 +6,7 @@ import (
 
 func convertYUVAction(c *cli.Context) error {
 	runtimePath := c.String("runtime")
-	generateOutFilePath, err := generate(runtimePath, c.String("file"))
+	generateOutFilePath, err := generateWithJIT(runtimePath, c.String("file"))
 	if err != nil {
 		return err
 	}
