+++ cli/genrun/grayscale.go
@@ -6,7 +6,7 @@ import (
 
 func grayscaleAction(c *cli.Context) error {
 	runtimePath := c.String("runtime")
-	generateOutFilePath, err := generate(runtimePath, c.String("file"))
+	generateOutFilePath, err := generateWithJIT(runtimePath, c.String("file"))
 	if err != nil {
 		return err
 	}
