+++ cli/genrun/laplacian.go
@@ -6,7 +6,7 @@ import (
 
 func laplacianAction(c *cli.Context) error {
 	runtimePath := c.String("runtime")
-	generateOutFilePath, err := generate(runtimePath, c.String("file"))
+	generateOutFilePath, err := generateWithJIT(runtimePath, c.String("file"))
 	if err != nil {
 		return err
 	}
